# By u210691
# Via u210691
* commit 'f6bbdf3dc70c42f57523b59c7f92776a7a6eba78':
  WZU-3111 bug fix
  WZU-3111 renamed and fixed bug
  WZU-3111 refactored to not accidentally delete not jenkins-slaves containers
  WZU-3111 Change nach Feedback von Christian und Roger
  WZU-3111 added and improved remove-jenkins-slave.sh

+++ _scripts/remove-container.sh
@@ -2,17 +2,26 @@
 containerid=$1
 externalfshome=/var/data/docker/container-ext-filesystems
 
-# stop container
-sudo docker stop $containerid
+if [ ! -z $containerid ]
+then
+	# stop container
+	sudo docker stop $containerid
+	if [ $? -eq 0 ]
+	then 
 
-# get fs id
-fsid=`sudo docker inspect -f='{{.Name}}' ${containerid} | awk -F"-" '{print $NF }'`
-
-# remove container
-sudo docker rm $containerid
-
-# remove fs
-if [[ ! -z "$fsid" ]]; then
-    sudo rm -rf $externalfshome/$fsid
+		# get fs id
+		number=`sudo docker inspect -f='{{.Name}}' ${containerid} | awk -F "-" '{ print $NF }'`
+		# remove container
+		sudo docker rm $containerid
+		# remove fs
+		if [ -d $externalfshome/$number ]
+		then
+			sudo rm -rf $externalfshome/$number
+		fi
+	else
+		echo "could not stop container with id ${containerid}"
+		exit 1
+	fi	
+else
+	echo "please provide a containerid"
 fi
-
